abstract out the working directory
/tmp/youtube-dl-playlist was hard-coded in several places.  Pull
them out into one variable definition as a start towards fixing
this.

A hardcoded path in /tmp is a security problem (predictable
temporary directory / symlink attacks on a shared system).  It's
also not ideal if you don't want to save stuff to your /tmp or /
filesystems (in my case, I have a separate partition for large
media).